refactor(SPV-1021): extract merkleroots as separate service
